Add files via upload
In this project I have created a machine learning model that classify whether the type of the sonar signal is rock or mine. It started from loading and looking at the sonar dataset, which has features based on multiple frequencies. After that, I did some data preprocessing and split the data into training set and testing set so as to not overfit it. I developed a predictive model using binary logistic regression and made predictions for which I calculated the accuracy metrics. The project deals with the techniques of machine learning (estimators) and how they apply to binary classification in problems related to analyzing sonar signals from mines.